+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Indentation from 3 to 4 spaces
+- Require placeholder only if `--placeholder-required` is used
 
 ## [1.0.0] - 2021-09-26